Made own VC for adding jams as a popover thingy..
